+++ src/ssl_mgr/config/renew_info.py
@@ -17,40 +17,33 @@ Default renew targets
 # pylint: disable=too-many-instance-attributes
 import random
 
+from .renew_func_from_dict import renew_func_from_dict
+from .renew_target import renew_target
+
 
 class RenewInfo:
     """
     When to renew
     """
     def __init__(self):
-        # renew target days remaining time to cert expiration
-        self.target_90: float = 30.0        # was renew_expire_days
-        self.target_60: float = 20.0
-        self.target_45: float = 10.0
-        self.target_10: float = 5.0
-        self.target_6: float = 2.0
-        self.target_2: float = 1.0
-        self.target_1: float = 0.5
-
-        # random variability days (0 means no variability)
-        self.rand_adj_90: float = 0.0       # was renew_expire_days_spread
-        self.rand_adj_60: float = 0.0
-        self.rand_adj_45: float = 0.0
-        self.rand_adj_10: float = 0.0
-        self.rand_adj_6: float = 0.0
-        self.rand_adj_2: float = 0.0
-        self.rand_adj_1: float = 0.0
+        #
+        # target function is a list of (orig_expiry, days_to_renew)
+        # rand_adj = list of (orig_expiry, random adjustment days)
+        # Lower bound fixed at (0, 0.25)
+        # Upper bound for expiry >= 300 is 20% of expiry
+        #
+        self.renew_func: list[tuple[float, float]] = []
+        self.rand_adj_func: list[tuple[float, float]] = []
 
     def from_dict(self, data_dict: dict[str, float]):
         """
         Extract data from dictionary
         """
-        for (k, v) in data_dict.items():
-            setattr(self, k, v)
+        renew_func_from_dict(data_dict, self.renew_func, self.rand_adj_func)
 
     def target(self, issue_days: float) -> tuple[float, float, float]:
         """
-        Return target in days to use for cer with original expiration
+        Return target in days to use for cert with original expiration
         of "issue_days"
 
         Input:
@@ -65,59 +58,7 @@ class RenewInfo:
         rand_adj_size: float = 0.0
         rand_adj_used: float = 0.0
 
-        if issue_days >= 89.0:
-            self.target_90 = _bounds_check(1.0, 89.0, self.target_90)
-            self.rand_adj_90 = _bounds_check(0.0, 5.0, self.rand_adj_90)
-
-            target = self.target_90
-            rand_adj_size = self.rand_adj_90
-
-        elif issue_days >= 59.0:
-            self.target_60 = _bounds_check(1.0, 59.0, self.target_60)
-            self.rand_adj_60 = _bounds_check(0.0, 3.0, self.rand_adj_60)
-
-            target = self.target_60
-            rand_adj_size = self.rand_adj_60
-
-        elif issue_days >= 44.0:
-            self.target_45 = _bounds_check(1.0, 44.0, self.target_45)
-            self.rand_adj_45 = _bounds_check(0.0, 2.0, self.rand_adj_45)
-
-            target = self.target_45
-            rand_adj_size = self.rand_adj_45
-
-        elif issue_days >= 9.0:
-            self.target_10 = _bounds_check(1.0, 9.0, self.target_10)
-            self.rand_adj_10 = _bounds_check(0.0, 1.0, self.rand_adj_10)
-
-            target = self.target_10
-            rand_adj_size = self.rand_adj_10
-
-        elif issue_days >= 5.0:
-            self.target_6 = _bounds_check(1.0, 5.0, self.target_6)
-            self.rand_adj_6 = _bounds_check(0.0, 1.0, self.rand_adj_6)
-
-            target = self.target_6
-            rand_adj_size = self.rand_adj_6
-
-        elif issue_days >= 1.9:
-            self.target_2 = _bounds_check(1.0, 1.9, self.target_2)
-            self.rand_adj_2 = _bounds_check(0.0, 0.5, self.rand_adj_2)
-
-            target = self.target_2
-            rand_adj_size = self.rand_adj_2
-
-        elif issue_days >= self.target_1:
-            self.target_1 = _bounds_check(0.1, 0.9, self.target_1)
-            self.rand_adj_1 = _bounds_check(0.0, 0.1, self.rand_adj_1)
-
-            target = self.target_1
-            rand_adj_size = self.rand_adj_1
-
-        else:
-            target = 0.25
-            rand_adj_size = 0.0
-
+        (target, rand_adj_size) = renew_target(self.renew_func, self.rand_adj_func, issue_days)
         if rand_adj_size > 0:
             rand_adj_used = _rand_adj(rand_adj_size)